refactor: legacy mvc 제거
