+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/ModelAndView.java
@@ -30,8 +30,4 @@ public class ModelAndView {
     public View getView() {
         return view;
     }
-
-    public String getViewName() {
-        return this.view.getViewName();
-    }
 }
