+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/View.java
@@ -6,6 +6,4 @@ import java.util.Map;
 
 public interface View {
     void render(Map<String, ?> model, HttpServletRequest request, HttpServletResponse response) throws Exception;
-
-    String getViewName();
 }
