+++ mvc/src/main/java/webmvc/org/springframework/web/servlet/view/JsonView.java
@@ -11,9 +11,4 @@ public class JsonView implements View {
     public void render(final Map<String, ?> model, final HttpServletRequest request, HttpServletResponse response)
             throws Exception {
     }
-
-    @Override
-    public String getViewName() {
-        return null;
-    }
 }
